refactor: simplify middleware and types
