+++ src/main/__tests__/middleware.test.ts
@@ -63,7 +63,7 @@ describe("the getForwardToRendererMiddlewareCreator function", () => {
 
     const alreadyForwardedAction = {
       type: "TEST_ACTION",
-      meta: { redial: { forwarded: true, source: "main", frameId: 0 } },
+      meta: { redial: { forwarded: true } },
     };
     expect(middleware(store)(next)(alreadyForwardedAction)).toEqual(alreadyForwardedAction);
 
@@ -71,7 +71,7 @@ describe("the getForwardToRendererMiddlewareCreator function", () => {
 
     const incrementAction = counterSlice.actions.increment();
     const result = middleware(store)(next)(incrementAction) as RedialAction;
-    expect(result.meta.redial).toEqual({ forwarded: true, source: "main", frameId: 0 });
+    expect(result.meta.redial).toEqual({ forwarded: true });
     expect(beforeSend).toHaveBeenCalledWith(result);
     expect(afterSend).toHaveBeenCalledWith(result);
     expect(mocks.send).toHaveBeenCalledWith(IpcChannel.FromMain, result);
@@ -88,7 +88,7 @@ describe("the getForwardToRendererMiddlewareCreator function", () => {
 
     const incrementAction = counterSlice.actions.increment();
     const result = middleware(store)(next)(incrementAction) as RedialAction;
-    expect(result.meta.redial).toEqual({ forwarded: true, source: "main", frameId: 0 });
+    expect(result.meta.redial).toEqual({ forwarded: true });
     expect(mocks.send).toHaveBeenCalledWith(IpcChannel.FromMain, result);
   });
 });
